Finish init project's test cases + Edit Documents (Add Test cases and answer question in reflection)

+++ TDDPratice/Utilities/OperationHelpers.cs
@@ -20,7 +20,7 @@ namespace TDDPratice.Utilities
 			throw new NotImplementedException();
 		}
 
-		public static long GetAverage(List<List<long>> input, int lineNumber)
+		public static double GetAverage(List<List<long>> input, int lineNumber)
 		{
 			throw new NotImplementedException();
 		}
